@@ -1,8 +1,6 @@
 module github.com/GustavoCaso/expensetrace
 
-go 1.23.0
-
-toolchain go1.23.7
+go 1.25.0
 
 require (
 	github.com/charmbracelet/bubbles v0.20.0
