@@ -1,7 +1,9 @@
 package db
 
 import (
+	"context"
 	"database/sql"
+
 	// import sqlite driver.
 	_ "github.com/mattn/go-sqlite3"
 )
@@ -13,7 +15,7 @@ func GetDB(dbsource string) (*sql.DB, error) {
 	}
 
 	// Enable foreign key constraints
-	_, err = db.Exec("PRAGMA foreign_keys = ON")
+	_, err = db.ExecContext(context.Background(), "PRAGMA foreign_keys = ON")
 	if err != nil {
 		return nil, err
 	}
